+++ scos_actions/actions/system_exit_action.py
