+++ scos_actions/discover/__init__.py
@@ -1,7 +1,7 @@
 from scos_actions.actions import action_classes
 from scos_actions.actions.logger import Logger
-from scos_actions.actions.RuntimeException import RuntimeException
-from scos_actions.actions.SystemExitAction import SystemExitAction
+from scos_actions.actions.runtime_exception import RuntimeException
+from scos_actions.actions.system_exit_action import SystemExitAction
 from scos_actions.actions.monitor_sigan import MonitorSignalAnalyzer
 from scos_actions.actions.sync_gps import SyncGps
 from scos_actions.discover.yaml import load_from_yaml
